+++ Code/dynamic.f90
@@ -270,9 +270,9 @@ CONTAINS
        DO nst=1,nstloc
           iq=isospin(globalindex(nst))
           CALL hpsi(iq,esf,psi(:,:,:,:,nst),ps1)
-          sp_energy(globalindex(nst))=overlap(psi(:,:,:,:,nst),ps1)
+          sp_energy(globalindex(nst))=REAL(overlap(psi(:,:,:,:,nst),ps1))
           sp_norm(globalindex(nst))=&
-               overlap(psi(:,:,:,:,nst),psi(:,:,:,:,nst))
+               REAL(overlap(psi(:,:,:,:,nst),psi(:,:,:,:,nst)))
        ENDDO
        CALL sp_properties
        IF(tmpi) THEN
